+++ test/all.dart
@@ -1,7 +1,15 @@
+import 'delegating_iterable_test.dart' as delegating_iterable_test;
+import 'delegating_list_test.dart' as delegating_list_test;
+import 'delegating_map_test.dart' as delegating_map_test;
 import 'iterable_test.dart' as iterable_test;
 import 'list_test.dart' as list_test;
+import 'map_test.dart' as map_test;
 
 void main() {
+  delegating_iterable_test.main();
+  delegating_list_test.main();
+  delegating_map_test.main();
   iterable_test.main();
   list_test.main();
+  map_test.main();
 }