+++ test/list_test.dart
@@ -1,8 +1,6 @@
 import 'package:superpower/superpower.dart';
 import 'package:test/test.dart';
 
-import 'common_list_iterable.dart';
-
 $List<int> get empty => $List<int>();
 
 void main() {
@@ -56,7 +54,12 @@ void main() {
   });
 
   test("test elementAtOrElse", () {
-    testElementAtOrElse(empty, $([0, 1, 2, 3]));
+    var elements = $([0, 1, 2, 3]);
+    expect(elements.elementAtOrElse(2, (_) => null), 2);
+    expect(elements.elementAtOrElse(4, (i) => i), 4);
+
+    expect(empty.elementAtOrElse(1, (i) => i), 1);
+    expect(empty.elementAtOrElse(0, (i) => i), 0);
   });
 
   test("test dropLast", () {